Fix for vaccination data page scraping (#3)
* Add: MOH case demographics data

* FIx: target directory missing

* Flat: latest data (2022-02-18T19:23:53.383Z)
{
  "date": "2022-02-18T19:23:53.383Z",
  "files": [
    {
      "name": "data/moh/covid-19-case-demographics/2022-02-17.html",
      "deltaBytes": 87751,
      "source": "https://www.health.govt.nz/covid-19-novel-coronavirus/covid-19-data-and-statistics/covid-19-case-demographics"
    }
  ]
}

* Add: Other MOH stats pages

* Flat: latest data (2022-02-18T23:11:32.175Z)
{
  "date": "2022-02-18T23:11:32.175Z",
  "files": [
    {
      "name": "data/moh/covid-19-current-cases/2022-02-17.html",
      "deltaBytes": 85877,
      "source": "https://www.health.govt.nz/covid-19-novel-coronavirus/covid-19-data-and-statistics/covid-19-current-cases"
    },
    {
      "name": "data/moh/covid-19-case-demographics/2022-02-17.html",
      "deltaBytes": 0,
      "source": "https://www.health.govt.nz/covid-19-novel-coronavirus/covid-19-data-and-statistics/covid-19-case-demographics"
    },
    {
      "name": "data/moh/covid-19-source-cases/2022-02-17.html",
      "deltaBytes": 85099,
      "source": "https://www.health.govt.nz/covid-19-novel-coronavirus/covid-19-data-and-statistics/covid-19-source-cases"
    },
    {
      "name": "data/moh/testing-covid-19/2022-02-17.html",
      "deltaBytes": 95382,
      "source": "https://www.health.govt.nz/covid-19-novel-coronavirus/covid-19-data-and-statistics/testing-covid-19"
    },
    {
      "name": "data/moh/covid-19-vaccine-data/Invalid Date.html",
      "deltaBytes": 0,
      "source": "https://www.health.govt.nz/covid-19-novel-coronavirus/covid-19-data-and-statistics/covid-19-vaccine-data"
    }
  ]
}

* Fix: Handle different date format in covid-19-vaccine-data.html

* Flat: latest data (2022-02-19T06:24:27.508Z)
{
  "date": "2022-02-19T06:24:27.508Z",
  "files": [
    {
      "name": "data/jhu-csse/time_series_covid19_deaths_nz.csv",
      "deltaBytes": 11,
      "source": "https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_deaths_global.csv"
    },
    {
      "name": "data/jhu-csse/time_series_covid19_confirmed_nz.csv",
      "deltaBytes": 14,
      "source": "https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_confirmed_global.csv"
    },
    {
      "name": "data/moh/covid-19-current-cases/2022-02-18.html",
      "deltaBytes": 85880,
      "source": "https://www.health.govt.nz/covid-19-novel-coronavirus/covid-19-data-and-statistics/covid-19-current-cases"
    },
    {
      "name": "data/moh/covid-19-case-demographics/2022-02-18.html",
      "deltaBytes": 87728,
      "source": "https://www.health.govt.nz/covid-19-novel-coronavirus/covid-19-data-and-statistics/covid-19-case-demographics"
    },
    {
      "name": "data/moh/covid-19-source-cases/2022-02-18.html",
      "deltaBytes": 85098,
      "source": "https://www.health.govt.nz/covid-19-novel-coronavirus/covid-19-data-and-statistics/covid-19-source-cases"
    },
    {
      "name": "data/moh/testing-covid-19/2022-02-18.html",
      "deltaBytes": 95383,
      "source": "https://www.health.govt.nz/covid-19-novel-coronavirus/covid-19-data-and-statistics/testing-covid-19"
    },
    {
      "name": "data/moh/covid-19-vaccine-data/2022-02-18.html",
      "deltaBytes": 177961,
      "source": "https://www.health.govt.nz/covid-19-novel-coronavirus/covid-19-data-and-statistics/covid-19-vaccine-data"
    }
  ]
}

* Flat: latest data (2022-02-19T06:37:34.169Z)
{
  "date": "2022-02-19T06:37:34.169Z",
  "files": [
    {
      "name": "data/moh/covid-19-case-demographics/2022-02-18.html",
      "deltaBytes": -76771,
      "source": "https://www.health.govt.nz/covid-19-novel-coronavirus/covid-19-data-and-statistics/covid-19-case-demographics"
    },
    {
      "name": "data/moh/covid-19-source-cases/2022-02-18.html",
      "deltaBytes": 0,
      "source": "https://www.health.govt.nz/covid-19-novel-coronavirus/covid-19-data-and-statistics/covid-19-source-cases"
    }
  ]
}

+++ postprocess/moh.ts
@@ -1,24 +1,50 @@
-/* This postprocessing script names the file based on the following string contained in the HTML.
+/* This postprocessing script names the file based on the a string like the following in the HTML.
+
   "All data on this page relates to cases recorded prior to 11:59 pm 17 February 2022."
+  
   The above example will become "2022-02-17.html" 
 */
+import { dirname } from 'https://deno.land/std@0.126.0/path/mod.ts';
 import dayjs from 'https://cdn.skypack.dev/dayjs';
+import customParseFormat from 'https://cdn.skypack.dev/dayjs/plugin/customParseFormat';
+dayjs.extend(customParseFormat);
 
 const inputFile = Deno.args[0];
 const html = Deno.readTextFileSync(inputFile);
+let dateRegex: RegExp;
 
-const regex =
-  /All data on this page relates to cases recorded prior to (?<date>.+)\./;
-const match = regex.exec(html);
+if (inputFile.endsWith('covid-19-current-cases.html')) {
+  dateRegex =
+    /All data on this page relates to cases recorded prior to (?<date>.+)\./;
+} else if (inputFile.endsWith('covid-19-case-demographics.html')) {
+  dateRegex =
+    /All data on this page relates to cases recorded prior to (?<date>.+)\./;
+} else if (inputFile.endsWith('covid-19-source-cases.html')) {
+  dateRegex =
+    /All data on this page relates to cases recorded prior to (?<date>.+)\./;
+} else if (inputFile.endsWith('testing-covid-19.html')) {
+  dateRegex =
+    /All data on this page relates to tests processed prior to (?<date>.+)\./;
+} else if (inputFile.endsWith('covid-19-vaccine-data.html')) {
+  dateRegex = /Data in this section is as at (?<date>.+) and is updated daily/;
+} else {
+  throw 'Unrecognised file: ' + inputFile;
+}
 
+const match = dateRegex.exec(html);
 if (!match?.groups?.date) {
-  throw 'Unable to work out what date the data relates to';
-  // TODO: Save to a fail directory
+  throw 'Unable to work out what date the data relates to: ' + inputFile;
 }
 
 const dateString = match.groups.date;
-const date = dayjs(dateString).format('YYYY-MM-DD');
+const date = dayjs(dateString, ['HH:mm a D MMMM YYYY', 'HH:mma D MMMM YYYY']);
 
-const outputFile = `./data/moh/covid-19-case-demographics/${date}.html`;
-Deno.writeTextFileSync(outputFile, html);
+if (!date.isValid()) {
+  throw 'Invalid date: ' + dateString;
+}
+
+const outputDir = dirname(inputFile);
+const outputFileName = `${date.format('YYYY-MM-DD')}.html`;
+
+Deno.writeTextFileSync(`./${outputDir}/${outputFileName}`, html);
 Deno.remove(inputFile);
